refactor : view 를 input 과 output 으로 나누고 리팩터링
